requirements.txt
rembg==2.0.67
onnxruntime==1.22.1
Pillow==11.2.1